==> magazines/01_mag_price_change.ipynb <==
## Данный ноутбук рассчитывает изменение цены для каждого тикера

import pandas as pd
from tqdm import tqdm_notebook as tqdm

### Загружаем цены акций

tickers = ['vtbr', 'sber', 'aflt', 'gazp', 'zill', 'aptk', 'mail', 'mtss',
       'yndx', 'tnse', 'qiwi', 'nlmk', 'irao', 'rosn', 'lnta', 'rual',
       'nmtp', 'gmkn', 'magn', 'mvid', 'mgnt', 'nvtk', 'tgkd', 'plzl',
       'five', 'kbtk', 'rtkm', 'kmaz', 'afks', 'akrn', 'lkoh', 'npof',
       'svav', 'usbn', 'tanl', 'rosb', 'tcsg', 'nknc', 'trcn', 'dvec',
       'moex', 'cbom']

p = pd.DataFrame()

# читаем цену 1 тикера 
for i in tickers:    
    try:
        p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep=',', dtype = {'<DATE>': 'str'})
        if p0.shape[1] == 1:
            p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep=';', dtype = {'<DATE>': 'str'})
        if p0.shape[1] == 1:
            p0 = pd.read_csv(f'/is/prices/{i.upper()}.csv', sep='|', dtype = {'<DATE>': 'str'})

        ls = []
        for c in p0.columns:
            ls.append(c.lower().replace('<', '').replace('>', ''))
        p0.columns = ls

        p = pd.concat([p, p0])
    except:
        continue
p.shape
    

p.head()

p.date.max()

#### Цены были получены из разных источников, поэтому требуется перевести даты цен в один вид

p.date = p.date.astype(str)
p['yea'] = p.date.apply(lambda x: x[:4])
p.head()

p1 = p[(p.yea!='2013')&(p.yea!='2014')&(p.yea!='2015')&(p.yea!='2016')&(p.yea!='2017')&(p.yea!='2018')&(p.yea!='2019')\
  &(p.yea!='2020')&(p.yea!='2021')]
p2 = p[(p.yea=='2013')|(p.yea=='2014')|(p.yea=='2015')|(p.yea=='2016')|(p.yea=='2017')|(p.yea=='2018')|(p.yea=='2019')\
  |(p.yea=='2020')|(p.yea=='2021')]

p1['d'] = p1.date.apply(lambda x: x[:2])
p1['m'] = p1.date.apply(lambda x: x[2:4])
p1['y'] = p1.date.apply(lambda x: '20' + x[4:6])
p1['date2'] = p1['y'] + p1['m'] + p1['d']
p1.head()

p1.date = p1.date2
p.columns

p1 = p1[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol']]
p1.head()

p2 = p2[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol']]
p2.head()

p = pd.concat([p1, p2])
p.head()

p.date = p.date.astype(float)

p.to_csv('/is/prices/mag_price.csv', sep='|', index = False)

p = pd.read_csv('/is/prices/mag_price.csv', sep='|')
p.head()

#### Устанавливает флаг, что данная запись была изначальной

p['init'] = 1
p.head()

#### Считаем изменение в ценах

p.shape

def lag(df, price0='close', price1='close', n=1, date='date'):
    
    '''
    Функция сортирует поле price по полю date по убыванию.
    Возвращает значение доходности в день начала окна.
    
    ''' 
    
    if price0 == price1 and n > 0:
        return (df.sort_values(date, ascending=False)[price1].shift(n) / \
                df.sort_values(date, ascending=False)[price0] - 1) * 100
    
    elif price0 == price1 and n == 0:
        return (df.sort_values(date, ascending=False)[price1] / \
                df.sort_values(date, ascending=False)[price0].shift(-1) - 1) * 100


for j in [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]:
    p[f'change_close_{j}'] = None
p.head()

p = p.sort_values(['ticker', 'date'], ascending=False)
pf1 = pd.DataFrame()
pf = pd.DataFrame()
for i in tqdm(p.ticker.unique(), position=0, leave=True):
    pf = p.loc[p.ticker == i]
    for j in [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]:
        pf[f'change_close_{j}'] = lag(df=pf, price0='close', price1='close', n=j)
    pf1 = pf1.append(pf)
pf1.head()


# проверка
pf1.groupby('ticker')['change_close_1'].sum().sort_values(ascending=False)[:30]

pf1.to_csv('/is/prices/mag_price.csv', sep='|', index = False)

### Добавляем даты из новостей

p = pf1
#p = pd.read_csv('/is/prices/price_vol_ret.csv', sep='|')
dates = pd.read_csv('/is/texts/mag/mag_news_dates.csv', sep='|')
p.ticker = p.ticker.str.lower()
dates.ticker = dates.ticker.str.lower()
p.date = p.date.astype(float)
dates.date = dates.date.astype(float)
p.head()

for i in p.columns:
    print(
    f'''p0['mean_{i}'] = pd.concat([p0['{i}'].bfill(), p0['{i}'].ffill()]).groupby(level=0).mean()'''
    )

p.head()

p1 = pd.DataFrame()
for i in tqdm(p.ticker.unique(), position=0, leave=True):
    p0 = p[p.ticker==i]
    dt = dates[dates.ticker==i]
    p0 = p0.merge(dt, how='outer', on=['date', 'ticker'])
    p0 = p0.sort_values(['date'])
    
    '''p0['open2'] = p0.open.fillna(method='bfill') # т.е. берётся след цена открытия
    p0['close2'] = p0.close.fillna(method='ffill') # т.е. берётся пред цена закрытия 
    
    p0['high2'] = p0.high.fillna(method='bfill') # т.е. берётся пред цена     
    p0['low2'] = p0.low.fillna(method='bfill') # т.е. берётся пред цена 
    
    p0['vol2'] = p0.vol.fillna(method='bfill') # т.е. берётся пред цена закрытия 
    '''
    
    p0['ticker'] = i

    p0['mean_open'] = pd.concat([p0['open'].bfill(), p0['open'].ffill()]).groupby(level=0).mean()
    p0['mean_high'] = pd.concat([p0['high'].bfill(), p0['high'].ffill()]).groupby(level=0).mean()
    p0['mean_low'] = pd.concat([p0['low'].bfill(), p0['low'].ffill()]).groupby(level=0).mean()
    p0['mean_close'] = pd.concat([p0['close'].bfill(), p0['close'].ffill()]).groupby(level=0).mean()
    p0['mean_vol'] = pd.concat([p0['vol'].bfill(), p0['vol'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_1'] = pd.concat([p0['change_close_1'].bfill(), p0['change_close_1'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_2'] = pd.concat([p0['change_close_2'].bfill(), p0['change_close_2'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_3'] = pd.concat([p0['change_close_3'].bfill(), p0['change_close_3'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_4'] = pd.concat([p0['change_close_4'].bfill(), p0['change_close_4'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_5'] = pd.concat([p0['change_close_5'].bfill(), p0['change_close_5'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_6'] = pd.concat([p0['change_close_6'].bfill(), p0['change_close_6'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_7'] = pd.concat([p0['change_close_7'].bfill(), p0['change_close_7'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_8'] = pd.concat([p0['change_close_8'].bfill(), p0['change_close_8'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_9'] = pd.concat([p0['change_close_9'].bfill(), p0['change_close_9'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_10'] = pd.concat([p0['change_close_10'].bfill(), p0['change_close_10'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_15'] = pd.concat([p0['change_close_15'].bfill(), p0['change_close_15'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_20'] = pd.concat([p0['change_close_20'].bfill(), p0['change_close_20'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_25'] = pd.concat([p0['change_close_25'].bfill(), p0['change_close_25'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_30'] = pd.concat([p0['change_close_30'].bfill(), p0['change_close_30'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_40'] = pd.concat([p0['change_close_40'].bfill(), p0['change_close_40'].ffill()]).groupby(level=0).mean()
    p0['mean_change_close_50'] = pd.concat([p0['change_close_50'].bfill(), p0['change_close_50'].ffill()]).groupby(level=0).mean()

    p0 = p0[(p0.mean_open.isna()!=True)&(p0.mean_close.isna()!=True)]
    p0 = p0.drop_duplicates()
    p1 = pd.concat([p0, p1])
p1.head()

p0 = p1[p1.open.isna()!=True]
p1 = p1[p1.open.isna()==True]
print(p0.shape, p1.shape)

for i in p1.columns:
    print(f'''p1.{i} = p1.mean_{i}''')

p1.open = p1.mean_open
p1.high = p1.mean_high
p1.low = p1.mean_low
p1.close = p1.mean_close
p1.vol = p1.mean_vol
p1.change_close_1 = p1.mean_change_close_1
p1.change_close_2 = p1.mean_change_close_2
p1.change_close_3 = p1.mean_change_close_3
p1.change_close_4 = p1.mean_change_close_4
p1.change_close_5 = p1.mean_change_close_5
p1.change_close_6 = p1.mean_change_close_6
p1.change_close_7 = p1.mean_change_close_7
p1.change_close_8 = p1.mean_change_close_8
p1.change_close_9 = p1.mean_change_close_9
p1.change_close_10 = p1.mean_change_close_10
p1.change_close_15 = p1.mean_change_close_15
p1.change_close_20 = p1.mean_change_close_20
p1.change_close_25 = p1.mean_change_close_25
p1.change_close_30 = p1.mean_change_close_30
p1.change_close_40 = p1.mean_change_close_40
p1.change_close_50 = p1.mean_change_close_50


p1.head()

p0.columns

p = pd.concat([p1[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol',
       'init', 'change_close_1', 'change_close_2', 'change_close_3',
       'change_close_4', 'change_close_5', 'change_close_6', 'change_close_7',
       'change_close_8', 'change_close_9', 'change_close_10',
       'change_close_15', 'change_close_20', 'change_close_25',
       'change_close_30', 'change_close_40', 'change_close_50']], \
               p0[['ticker', 'per', 'date', 'time', 'open', 'high', 'low', 'close', 'vol',
       'init', 'change_close_1', 'change_close_2', 'change_close_3',
       'change_close_4', 'change_close_5', 'change_close_6', 'change_close_7',
       'change_close_8', 'change_close_9', 'change_close_10',
       'change_close_15', 'change_close_20', 'change_close_25',
       'change_close_30', 'change_close_40', 'change_close_50']]])
print(p.shape)
p.head()

#### Удаляем новости, которые вышли после крайней даты цен

p[p.init==1].groupby('ticker').date.min().sort_values()

p.groupby('ticker').date.min().sort_values()

df_temp = pd.DataFrame()
for ticker in p.ticker.unique():
    min_dt = p[(p.init==1) & (p.ticker==ticker)].date.min()
    max_dt = p[(p.init==1) & (p.ticker==ticker)].date.max()
    temp = p[(p.ticker==ticker) & (p.date >= min_dt) & (p.date <= max_dt)]
    df_temp = pd.concat([df_temp, temp])
df_temp.shape

p[p.init==1].groupby('ticker').date.min().sort_values() == p.groupby('ticker').date.min().sort_values()

p[p.init==1].groupby('ticker').date.max().sort_values() == p.groupby('ticker').date.max().sort_values()

df_temp[df_temp.init==1].groupby('ticker').date.min().sort_values() == df_temp.groupby('ticker').date.min().sort_values()

df_temp[df_temp.init==1].groupby('ticker').date.max().sort_values() == df_temp.groupby('ticker').date.max().sort_values()

df_temp.head()

df_temp[['init']] = df_temp[['init']].fillna(value=0)
df_temp.head()

print(df_temp.shape)
df_temp = df_temp.drop_duplicates()
print(df_temp.shape)

df_temp.to_csv('/is/prices/mag_price_change.csv', sep='|', index = False)
