==> magazines/03_mag_label_ticker_sentiment.ipynb <==
## В данном ноутбуке происходит labelling акций, join сентимента и цен

import pandas as pd
import numpy as np
from tqdm.notebook import tqdm

### Сентимент

s = pd.read_csv('/is/texts/mag/mag_text_lower_tic_all_flag_all_dt_e_reg_stop_stem_len.csv', sep='|')
s.head()

print(s.shape)
s = s[(s.text_reg_stop_stem != '')]
print(s.shape)
s = s[(s.text_reg_stop_stem.isna() != True)]
print(s.shape)
s = s.drop_duplicates()
print(s.shape)

s['msg_init'] = 1
s.head()

### Цена

p = pd.read_csv('/is/prices/mag_price_change.csv', sep='|')
p.shape

p.head()

s.date = s.date.astype(float)
p.date = p.date.astype(float)
s.ticker = s.ticker.str.lower()
p.ticker = p.ticker.str.lower()

print(s.shape, p.shape)

p.rename(columns = {'init': 'price_init'}, inplace=True)
p.head()

ps = p.merge(s, how='inner', on=['date', 'ticker'])
ps.to_csv('/is/prices/mag_price_sentiment_join.csv', sep='|', index = False)
print(ps.shape)
ps.head()

print(ps.shape)
ps = ps[ps.ticker.isin(p.ticker.unique())]
print(ps.shape)

for i in ps.columns: print(f"'{i}', ", sep='', end='')

ps.to_csv('/is/prices/mag_price_sentiment_join.csv', sep='|', index = False)

def labelling(df, price, field, n=0):
#исправить функцию чтобы она по массиву брала
    
    '''
    Функция определяет сентимент поля price
    и указывает в новом поле field:
    1 для положительного фона,
    0 для негативного фона,
    -1 для нейтрального фона.
    Фон определяется на основе уровня изменения цены - n.
    
    '''
    df[field] = None
    
    df[field][df[price] >= (n)] = 1
    df[field][df[price] <= (-n)] = -1
    df[field][(df[price] > (-n)) & (df[price] < (n))] = 0


np.linspace(1, 10, num=10)

yields = [0.05, 0.1, 0.2, 0.3, 0.5, 0.8, 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10.]
lags = [1,2,3,4,5,6,7,8,9,10, 15, 20, 25, 30, 40, 50]
for lag in tqdm(lags, position=0, leave=True): 
    for income in tqdm(yields, position=0, leave=True):
        labelling(df=ps, price=f'change_close_{lag}', field=f'label_{lag}_{income}', n=income)

ps.to_csv('/is/prices/mag_join.csv', sep='|', index = False)
ps.head()

ps.drop(columns=['per', 'time', 'open', 'high', 'low', 
       'text_reg_stop', 'msg_init'], inplace=True)

ps.to_csv('/is/prices/mag_join_lab.csv', sep='|', index = False)

ps.head()
