==> magazines/04_mag_NB.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report
from tqdm.notebook import tqdm
from sklearn.metrics import mean_squared_error
from sklearn import datasets, linear_model
from sklearn.feature_extraction import DictVectorizer

from sklearn.model_selection import GridSearchCV
from sklearn.svm import SVC, SVR
from sklearn.model_selection import KFold
import sklearn
import warnings
warnings.filterwarnings('ignore')

import statsmodels.api as sm
import pandas as pd

f = pd.read_csv('/is/prices/mag_join_lab.csv', sep='|')

print(f.shape)
b = pd.read_csv('/is/prices/mag_bh.csv', sep='|')


f = f.merge(b, how='inner', on='ticker')
f['obs'] = 1
f = f.sort_values('date', ascending=False)
print(f.shape)

j_neut = 0.8
j_pos = 7.0
j_neg = 6.0
i = 10
label_neut = f'label_{i}_{j_neut}'
label_pos = f'label_{i}_{j_pos}'
label_neg = f'label_{i}_{j_neg}'
print('neut: ', f[label_neut][f[label_neut]==0].shape)
print('neg: ', f[label_neg][f[label_neg]==-1].shape)
print('pos: ', f[label_pos][f[label_pos]==1].shape)

lagi = [1]
j_neuts = [0.2]
j_poss = [2.0]
j_negs = [2.0]


lagi = [2]
j_neuts = [0.5]
j_poss = [3.0]
j_negs=[3.0]


lagi = [5]
j_neuts = [0.5]
j_poss = [5.0]
j_negs = [5.0]


lagi = [10]
j_neuts = [0.8]
j_poss = [7.0]
j_negs = [6.0]


6 * 3 * 41

#### Финальный репорт

from sklearn import metrics
field = 'text_reg_stop_stem'
out = pd.DataFrame()

min_dfs = [5, 10, 15, 25, 35, 50] 
max_dfs = [0.5]
ngram_range_maxs = [1]
lagi = [1]
j_neuts = [0.2]
j_poss = [4.0]
j_negs = [4.0]
days = [1]

counter = 0

t = 4
for i in lagi:
    # split
    for ticker in  f.ticker.unique(): #['sber']: #
        print(ticker)

        for j_neut in j_neuts:
            for j_pos in j_poss:
                for j_neg in j_negs:
                    label_neut = f'label_{i}_{j_neut}'
                    label_pos = f'label_{i}_{j_pos}'
                    label_neg = f'label_{i}_{j_neg}'
                    for day in days:
                        try:

                            ps = f[[label_neut, label_pos, label_neg, field]][(f.ticker!=ticker)] 
                            ps = ps.loc[:,~ps.columns.duplicated()]
                            if t == 0:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neg]==-1)].sample(frac=1, random_state=42)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neg:'sentiment'}), \
                                               ]).sample(frac=1, random_state=44)

                            else:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neut]==0)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)][[label_pos, field]].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neut]==0)][[label_neut, field]].sample(frac=1, random_state=42)
                                ps3 = ps[(ps[label_neg]==-1)][[label_neg, field]].sample(frac=1, random_state=43)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neut:'sentiment'}), \
                                                ps3[:min_df].rename(columns={label_neg:'sentiment'}) \
                                               ]).sample(frac=1, random_state=44)


                            x_train = df[field]
                            y_train = df['sentiment'] 

                        # fit
                            for ngram_range_max in ngram_range_maxs:
                                for min_df in min_dfs:
                                    for max_df in max_dfs:

                                        tf_idf = TfidfVectorizer(min_df=min_df, \
                                                                 max_df=max_df, \
                                                                 ngram_range=(1, ngram_range_max))

                                        train_features = tf_idf.fit_transform(x_train)


                                        model = MultinomialNB(alpha=0.1)
                                        model.fit(train_features, y_train)
                                        y_train_pred = model.predict(train_features)
        # В идеале волатильность взять за эти же лаги наверное
                                        df_temp = pd.DataFrame(f[(f.ticker==ticker)]\
                                                               [[field, 'obs', 'date',\
                                                                 f'change_close_{day}', 'bh', 'close']]).\
            reset_index().drop('index', axis=1)

                                        y_pred = model.predict(tf_idf.transform(df_temp[field]))

                                        df_temp['sentiment'] = y_pred

                                        df_temp['neg'] = 0
                                        df_temp['neg'][df_temp.sentiment==-1] = 1

                                        df_temp['pos'] = 0
                                        df_temp['pos'][df_temp.sentiment==1] = 1

                                        df_temp['neut'] = 0
                                        df_temp['neut'][df_temp.sentiment==0] = 1

                                        df_temp = df_temp.drop_duplicates().\
                                        groupby(['date', f'change_close_{day}', 'bh', 'close']).sum().reset_index()

        # Делаем сум роллинг окно для суммирования сентиментов для периодов больше 1 дня 
        # Нужно удостовериться в сортировке должно быть по возрастанию даты

                                        df_temp[f'close_lag_1'] = df_temp.close.shift(periods=1)
                                        #df_temp[f'close_change_lag_1'] = df_temp.close.shift(periods=1)
                                        df_temp = df_temp.sort_values('date', ascending=True)

                                        df_temp['neg'] = df_temp['neg'].rolling(day).sum()
                                        df_temp['pos'] = df_temp['pos'].rolling(day).sum()
                                        df_temp['neut'] = df_temp['neut'].rolling(day).sum()

                                        df_temp.dropna(subset=['neg', 'pos', 'neut', 'close_lag_1'], inplace=True)
                        # просто сумма
                                        df_temp['sum_sentiment'] = 0
                                        df_temp['sum_sentiment'][(df_temp.pos>df_temp.neg)&(df_temp.pos>=df_temp.neut)] = 1
                                        df_temp['sum_sentiment'][(df_temp.pos<df_temp.neg)&(df_temp.neg>=df_temp.neut)] = -1

        # дивергенция c добавлением 
                                        df_temp['diver'] = ((df_temp.pos - df_temp.neg)**2 \
                                                                  / (df_temp.pos + df_temp.neg)**2)**(1/2)

                                        #df_temp.reset_index(level=[f'change_close_{day}', 'date', 'bh'], inplace=True)
                                        df_temp['clear_return'] = df_temp.sum_sentiment * df_temp[f'change_close_{day}']
                                        df_temp['diver_return'] = df_temp.clear_return * df_temp.diver                

                                        res = dict()
                                        res['counter'] = [counter]
                                        res['ticker'] = [ticker]
                                        res['date'] = [day]

                                        res['min_df'] = [min_df]

                                        res['j_neut'] = [j_neut]
                                        res['j_pos'] = [j_pos] 
                                        res['j_neg'] = [j_neg] 

                                        res['train_size'] = [y_train.shape[0]]
                                        res['test_size'] = [y_pred.shape[0]]

                                        res['obs'] = \
                                        [(df_temp.obs[df_temp.clear_return.isna()==False]).sum()]
                                        res['bh'] = df_temp.bh.max()
                                        res['original_return_isna_all'] = \
                                        [(df_temp[f'change_close_{day}']\
                                          [df_temp[f'change_close_{day}'].isna()==False]).sum()]


                                        res['model_return_isna_all'] = \
                                        [(df_temp.clear_return[df_temp.clear_return.isna()==False]).sum()]

                                        res['model_return_isna_short'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_return_isna_long'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]


                                        res['model_diver_return_isna_all'] = \
                                        [(df_temp.diver_return[df_temp.diver_return.isna()==False]).sum()]

                                        res['model_diver_return_isna_short'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_diver_return_isna_long'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]






                                        res['train_accuracy'] = [metrics.accuracy_score(y_train_pred, y_train)]


                                        res['train_precision_macro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='macro')]


                                        res['train_precision_micro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='micro')]


                                        res['train_recall_micro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='micro')]


                                        res['train_recall_macro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='macro')]

                                        res['train_f1_micro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='micro')]


                                        res['train_f1_macro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='macro')]



                                        x = df_temp[['sum_sentiment', 'close_lag_1']]
                                        x.sum_sentiment = x.sum_sentiment.astype(str)

                                        x = pd.get_dummies(x, drop_first=True)
                                        y = df_temp.close
                                        x2 = sm.add_constant(x)
                                        lm = sm.OLS(y, x2)
                                        res_lm = lm.fit()

                                        res['const'] = res_lm.params[0]
                                        res[f'{x.columns[0]}_coef'] = res_lm.params[1]
                                        res[f'{x.columns[1]}_coef'] = res_lm.params[2]
                                        res[f'{x.columns[2]}_coef'] = res_lm.params[3]

                                        res['const_p'] = res_lm.pvalues[0]
                                        res[f'{x.columns[0]}_p'] = res_lm.pvalues[1]
                                        res[f'{x.columns[1]}_p'] = res_lm.pvalues[2]
                                        res[f'{x.columns[2]}_p'] = res_lm.pvalues[3]
                                        res['r2'] = res_lm.rsquared
                                        res['nobs'] = res_lm.nobs
                                        counter += 1
                                        print(counter, end='\r')


                                        res = pd.DataFrame.from_dict(res)

                                        out = out.append(pd.DataFrame(res))
                                        if counter % 100 == 0: 
                                            out.to_excel(f'/is/res/nb/mag_nb_model_ind_final14.xlsx', index=False)

                        except:
                            continue
    

                    


out.to_excel(f'/is/res/nb/mag_nb_model_ind_final14.xlsx', index=False)
out

from sklearn import metrics
field = 'text_reg_stop_stem'
out = pd.DataFrame()

min_dfs = [5, 10, 15, 25, 35, 50] 
max_dfs = [0.5]
ngram_range_maxs = [1]
lagi = [5]
j_neuts = [0.5]
j_poss = [5.0]
j_negs = [5.0]

days = [1]

counter = 0

t = 4
for i in lagi:
    # split
    for ticker in  f.ticker.unique(): #['sber']: #
        print(ticker)

        for j_neut in j_neuts:
            for j_pos in j_poss:
                for j_neg in j_negs:
                    label_neut = f'label_{i}_{j_neut}'
                    label_pos = f'label_{i}_{j_pos}'
                    label_neg = f'label_{i}_{j_neg}'
                    for day in days:
                        try:

                            ps = f[[label_neut, label_pos, label_neg, field]][(f.ticker!=ticker)] 
                            ps = ps.loc[:,~ps.columns.duplicated()]
                            if t == 0:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neg]==-1)].sample(frac=1, random_state=42)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neg:'sentiment'}), \
                                               ]).sample(frac=1, random_state=44)

                            else:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neut]==0)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)][[label_pos, field]].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neut]==0)][[label_neut, field]].sample(frac=1, random_state=42)
                                ps3 = ps[(ps[label_neg]==-1)][[label_neg, field]].sample(frac=1, random_state=43)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neut:'sentiment'}), \
                                                ps3[:min_df].rename(columns={label_neg:'sentiment'}) \
                                               ]).sample(frac=1, random_state=44)


                            x_train = df[field]
                            y_train = df['sentiment'] 

                        # fit
                            for ngram_range_max in ngram_range_maxs:
                                for min_df in min_dfs:
                                    for max_df in max_dfs:

                                        tf_idf = TfidfVectorizer(min_df=min_df, \
                                                                 max_df=max_df, \
                                                                 ngram_range=(1, ngram_range_max))

                                        train_features = tf_idf.fit_transform(x_train)


                                        model = MultinomialNB(alpha=0.1)
                                        model.fit(train_features, y_train)
                                        y_train_pred = model.predict(train_features)
        # В идеале волатильность взять за эти же лаги наверное
                                        df_temp = pd.DataFrame(f[(f.ticker==ticker)]\
                                                               [[field, 'obs', 'date',\
                                                                 f'change_close_{day}', 'bh', 'close']]).\
            reset_index().drop('index', axis=1)

                                        y_pred = model.predict(tf_idf.transform(df_temp[field]))

                                        df_temp['sentiment'] = y_pred

                                        df_temp['neg'] = 0
                                        df_temp['neg'][df_temp.sentiment==-1] = 1

                                        df_temp['pos'] = 0
                                        df_temp['pos'][df_temp.sentiment==1] = 1

                                        df_temp['neut'] = 0
                                        df_temp['neut'][df_temp.sentiment==0] = 1

                                        df_temp = df_temp.drop_duplicates().\
                                        groupby(['date', f'change_close_{day}', 'bh', 'close']).sum().reset_index()

        # Делаем сум роллинг окно для суммирования сентиментов для периодов больше 1 дня 
        # Нужно удостовериться в сортировке должно быть по возрастанию даты

                                        df_temp[f'close_lag_1'] = df_temp.close.shift(periods=1)
                                        #df_temp[f'close_change_lag_1'] = df_temp.close.shift(periods=1)
                                        df_temp = df_temp.sort_values('date', ascending=True)

                                        df_temp['neg'] = df_temp['neg'].rolling(day).sum()
                                        df_temp['pos'] = df_temp['pos'].rolling(day).sum()
                                        df_temp['neut'] = df_temp['neut'].rolling(day).sum()

                                        df_temp.dropna(subset=['neg', 'pos', 'neut', 'close_lag_1'], inplace=True)
                        # просто сумма
                                        df_temp['sum_sentiment'] = 0
                                        df_temp['sum_sentiment'][(df_temp.pos>df_temp.neg)&(df_temp.pos>=df_temp.neut)] = 1
                                        df_temp['sum_sentiment'][(df_temp.pos<df_temp.neg)&(df_temp.neg>=df_temp.neut)] = -1

        # дивергенция c добавлением 
                                        df_temp['diver'] = ((df_temp.pos - df_temp.neg)**2 \
                                                                  / (df_temp.pos + df_temp.neg)**2)**(1/2)

                                        #df_temp.reset_index(level=[f'change_close_{day}', 'date', 'bh'], inplace=True)
                                        df_temp['clear_return'] = df_temp.sum_sentiment * df_temp[f'change_close_{day}']
                                        df_temp['diver_return'] = df_temp.clear_return * df_temp.diver                

                                        res = dict()
                                        res['counter'] = [counter]
                                        res['ticker'] = [ticker]
                                        res['date'] = [day]

                                        res['min_df'] = [min_df]

                                        res['j_neut'] = [j_neut]
                                        res['j_pos'] = [j_pos] 
                                        res['j_neg'] = [j_neg] 

                                        res['train_size'] = [y_train.shape[0]]
                                        res['test_size'] = [y_pred.shape[0]]

                                        res['obs'] = \
                                        [(df_temp.obs[df_temp.clear_return.isna()==False]).sum()]
                                        res['bh'] = df_temp.bh.max()
                                        res['original_return_isna_all'] = \
                                        [(df_temp[f'change_close_{day}']\
                                          [df_temp[f'change_close_{day}'].isna()==False]).sum()]


                                        res['model_return_isna_all'] = \
                                        [(df_temp.clear_return[df_temp.clear_return.isna()==False]).sum()]

                                        res['model_return_isna_short'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_return_isna_long'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]


                                        res['model_diver_return_isna_all'] = \
                                        [(df_temp.diver_return[df_temp.diver_return.isna()==False]).sum()]

                                        res['model_diver_return_isna_short'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_diver_return_isna_long'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]






                                        res['train_accuracy'] = [metrics.accuracy_score(y_train_pred, y_train)]


                                        res['train_precision_macro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='macro')]


                                        res['train_precision_micro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='micro')]


                                        res['train_recall_micro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='micro')]


                                        res['train_recall_macro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='macro')]

                                        res['train_f1_micro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='micro')]


                                        res['train_f1_macro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='macro')]



                                        x = df_temp[['sum_sentiment', 'close_lag_1']]
                                        x.sum_sentiment = x.sum_sentiment.astype(str)

                                        x = pd.get_dummies(x, drop_first=True)
                                        y = df_temp.close
                                        x2 = sm.add_constant(x)
                                        lm = sm.OLS(y, x2)
                                        res_lm = lm.fit()

                                        res['const'] = res_lm.params[0]
                                        res[f'{x.columns[0]}_coef'] = res_lm.params[1]
                                        res[f'{x.columns[1]}_coef'] = res_lm.params[2]
                                        res[f'{x.columns[2]}_coef'] = res_lm.params[3]

                                        res['const_p'] = res_lm.pvalues[0]
                                        res[f'{x.columns[0]}_p'] = res_lm.pvalues[1]
                                        res[f'{x.columns[1]}_p'] = res_lm.pvalues[2]
                                        res[f'{x.columns[2]}_p'] = res_lm.pvalues[3]
                                        res['r2'] = res_lm.rsquared
                                        res['nobs'] = res_lm.nobs
                                        counter += 1
                                        print(counter, end='\r')


                                        res = pd.DataFrame.from_dict(res)

                                        out = out.append(pd.DataFrame(res))
                                        if counter % 100 == 0: 
                                            out.to_excel(f'/is/res/nb/mag_nb_model_ind_final5.xlsx', index=False)

                        except:
                            continue
    

                    


out.to_excel(f'/is/res/nb/mag_nb_model_ind_final5.xlsx', index=False)
out

from sklearn import metrics
field = 'text_reg_stop_stem'
out = pd.DataFrame()

min_dfs = [5, 10, 15, 25, 35, 50] 
max_dfs = [0.5]
ngram_range_maxs = [1]
lagi = [2]
j_neuts = [0.5]
j_poss = [3.0]
j_negs=[3.0]
days = [1]

counter = 0

t = 4
for i in lagi:
    # split
    for ticker in  f.ticker.unique(): #['sber']: #
        print(ticker)

        for j_neut in j_neuts:
            for j_pos in j_poss:
                for j_neg in j_negs:
                    label_neut = f'label_{i}_{j_neut}'
                    label_pos = f'label_{i}_{j_pos}'
                    label_neg = f'label_{i}_{j_neg}'
                    for day in days:
                        try:

                            ps = f[[label_neut, label_pos, label_neg, field]][(f.ticker!=ticker)] 
                            ps = ps.loc[:,~ps.columns.duplicated()]
                            if t == 0:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neg]==-1)].sample(frac=1, random_state=42)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neg:'sentiment'}), \
                                               ]).sample(frac=1, random_state=44)

                            else:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neut]==0)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)][[label_pos, field]].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neut]==0)][[label_neut, field]].sample(frac=1, random_state=42)
                                ps3 = ps[(ps[label_neg]==-1)][[label_neg, field]].sample(frac=1, random_state=43)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neut:'sentiment'}), \
                                                ps3[:min_df].rename(columns={label_neg:'sentiment'}) \
                                               ]).sample(frac=1, random_state=44)


                            x_train = df[field]
                            y_train = df['sentiment'] 

                        # fit
                            for ngram_range_max in ngram_range_maxs:
                                for min_df in min_dfs:
                                    for max_df in max_dfs:

                                        tf_idf = TfidfVectorizer(min_df=min_df, \
                                                                 max_df=max_df, \
                                                                 ngram_range=(1, ngram_range_max))

                                        train_features = tf_idf.fit_transform(x_train)


                                        model = MultinomialNB(alpha=0.1)
                                        model.fit(train_features, y_train)
                                        y_train_pred = model.predict(train_features)
        # В идеале волатильность взять за эти же лаги наверное
                                        df_temp = pd.DataFrame(f[(f.ticker==ticker)]\
                                                               [[field, 'obs', 'date',\
                                                                 f'change_close_{day}', 'bh', 'close']]).\
            reset_index().drop('index', axis=1)

                                        y_pred = model.predict(tf_idf.transform(df_temp[field]))

                                        df_temp['sentiment'] = y_pred

                                        df_temp['neg'] = 0
                                        df_temp['neg'][df_temp.sentiment==-1] = 1

                                        df_temp['pos'] = 0
                                        df_temp['pos'][df_temp.sentiment==1] = 1

                                        df_temp['neut'] = 0
                                        df_temp['neut'][df_temp.sentiment==0] = 1

                                        df_temp = df_temp.drop_duplicates().\
                                        groupby(['date', f'change_close_{day}', 'bh', 'close']).sum().reset_index()

        # Делаем сум роллинг окно для суммирования сентиментов для периодов больше 1 дня 
        # Нужно удостовериться в сортировке должно быть по возрастанию даты

                                        df_temp[f'close_lag_1'] = df_temp.close.shift(periods=1)
                                        #df_temp[f'close_change_lag_1'] = df_temp.close.shift(periods=1)
                                        df_temp = df_temp.sort_values('date', ascending=True)

                                        df_temp['neg'] = df_temp['neg'].rolling(day).sum()
                                        df_temp['pos'] = df_temp['pos'].rolling(day).sum()
                                        df_temp['neut'] = df_temp['neut'].rolling(day).sum()

                                        df_temp.dropna(subset=['neg', 'pos', 'neut', 'close_lag_1'], inplace=True)
                        # просто сумма
                                        df_temp['sum_sentiment'] = 0
                                        df_temp['sum_sentiment'][(df_temp.pos>df_temp.neg)&(df_temp.pos>=df_temp.neut)] = 1
                                        df_temp['sum_sentiment'][(df_temp.pos<df_temp.neg)&(df_temp.neg>=df_temp.neut)] = -1

        # дивергенция c добавлением 
                                        df_temp['diver'] = ((df_temp.pos - df_temp.neg)**2 \
                                                                  / (df_temp.pos + df_temp.neg)**2)**(1/2)

                                        #df_temp.reset_index(level=[f'change_close_{day}', 'date', 'bh'], inplace=True)
                                        df_temp['clear_return'] = df_temp.sum_sentiment * df_temp[f'change_close_{day}']
                                        df_temp['diver_return'] = df_temp.clear_return * df_temp.diver                

                                        res = dict()
                                        res['counter'] = [counter]
                                        res['ticker'] = [ticker]
                                        res['date'] = [day]

                                        res['min_df'] = [min_df]

                                        res['j_neut'] = [j_neut]
                                        res['j_pos'] = [j_pos] 
                                        res['j_neg'] = [j_neg] 

                                        res['train_size'] = [y_train.shape[0]]
                                        res['test_size'] = [y_pred.shape[0]]

                                        res['obs'] = \
                                        [(df_temp.obs[df_temp.clear_return.isna()==False]).sum()]
                                        res['bh'] = df_temp.bh.max()
                                        res['original_return_isna_all'] = \
                                        [(df_temp[f'change_close_{day}']\
                                          [df_temp[f'change_close_{day}'].isna()==False]).sum()]


                                        res['model_return_isna_all'] = \
                                        [(df_temp.clear_return[df_temp.clear_return.isna()==False]).sum()]

                                        res['model_return_isna_short'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_return_isna_long'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]


                                        res['model_diver_return_isna_all'] = \
                                        [(df_temp.diver_return[df_temp.diver_return.isna()==False]).sum()]

                                        res['model_diver_return_isna_short'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_diver_return_isna_long'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]






                                        res['train_accuracy'] = [metrics.accuracy_score(y_train_pred, y_train)]


                                        res['train_precision_macro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='macro')]


                                        res['train_precision_micro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='micro')]


                                        res['train_recall_micro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='micro')]


                                        res['train_recall_macro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='macro')]

                                        res['train_f1_micro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='micro')]


                                        res['train_f1_macro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='macro')]



                                        x = df_temp[['sum_sentiment', 'close_lag_1']]
                                        x.sum_sentiment = x.sum_sentiment.astype(str)

                                        x = pd.get_dummies(x, drop_first=True)
                                        y = df_temp.close
                                        x2 = sm.add_constant(x)
                                        lm = sm.OLS(y, x2)
                                        res_lm = lm.fit()

                                        res['const'] = res_lm.params[0]
                                        res[f'{x.columns[0]}_coef'] = res_lm.params[1]
                                        res[f'{x.columns[1]}_coef'] = res_lm.params[2]
                                        res[f'{x.columns[2]}_coef'] = res_lm.params[3]

                                        res['const_p'] = res_lm.pvalues[0]
                                        res[f'{x.columns[0]}_p'] = res_lm.pvalues[1]
                                        res[f'{x.columns[1]}_p'] = res_lm.pvalues[2]
                                        res[f'{x.columns[2]}_p'] = res_lm.pvalues[3]
                                        res['r2'] = res_lm.rsquared
                                        res['nobs'] = res_lm.nobs
                                        counter += 1
                                        print(counter, end='\r')


                                        res = pd.DataFrame.from_dict(res)

                                        out = out.append(pd.DataFrame(res))
                                        if counter % 100 == 0: 
                                            out.to_excel(f'/is/res/nb/mag_nb_model_ind_final2.xlsx', index=False)

                        except:
                            continue
    

                    


out.to_excel(f'/is/res/nb/mag_nb_model_ind_final2.xlsx', index=False)
out

from sklearn import metrics
field = 'text_reg_stop_stem'
out = pd.DataFrame()

min_dfs = [5, 10, 15, 25, 35, 50] 
max_dfs = [0.5]
ngram_range_maxs = [1]

lagi = [1]
j_neuts = [0.2]
j_poss = [2.0]
j_negs = [2.0]

days = [1]

counter = 0

t = 4
for i in lagi:
    # split
    for ticker in  f.ticker.unique(): #['sber']: #
        print(ticker)

        for j_neut in j_neuts:
            for j_pos in j_poss:
                for j_neg in j_negs:
                    label_neut = f'label_{i}_{j_neut}'
                    label_pos = f'label_{i}_{j_pos}'
                    label_neg = f'label_{i}_{j_neg}'
                    for day in days:
                        try:

                            ps = f[[label_neut, label_pos, label_neg, field]][(f.ticker!=ticker)] 
                            ps = ps.loc[:,~ps.columns.duplicated()]
                            if t == 0:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neg]==-1)].sample(frac=1, random_state=42)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neg:'sentiment'}), \
                                               ]).sample(frac=1, random_state=44)

                            else:
                                min_df = min(ps[(ps[label_pos]==1)].shape[0], \
                                             ps[(ps[label_neut]==0)].shape[0], \
                                             ps[(ps[label_neg]==-1)].shape[0])

                                ps1 = ps[(ps[label_pos]==1)][[label_pos, field]].sample(frac=1, random_state=41)
                                ps2 = ps[(ps[label_neut]==0)][[label_neut, field]].sample(frac=1, random_state=42)
                                ps3 = ps[(ps[label_neg]==-1)][[label_neg, field]].sample(frac=1, random_state=43)

                                df = pd.concat([ps1[:min_df].rename(columns={label_pos:'sentiment'}), \
                                                ps2[:min_df].rename(columns={label_neut:'sentiment'}), \
                                                ps3[:min_df].rename(columns={label_neg:'sentiment'}) \
                                               ]).sample(frac=1, random_state=44)


                            x_train = df[field]
                            y_train = df['sentiment'] 

                        # fit
                            for ngram_range_max in ngram_range_maxs:
                                for min_df in min_dfs:
                                    for max_df in max_dfs:

                                        tf_idf = TfidfVectorizer(min_df=min_df, \
                                                                 max_df=max_df, \
                                                                 ngram_range=(1, ngram_range_max))

                                        train_features = tf_idf.fit_transform(x_train)


                                        model = MultinomialNB(alpha=0.1)
                                        model.fit(train_features, y_train)
                                        y_train_pred = model.predict(train_features)
        # В идеале волатильность взять за эти же лаги наверное
                                        df_temp = pd.DataFrame(f[(f.ticker==ticker)]\
                                                               [[field, 'obs', 'date',\
                                                                 f'change_close_{day}', 'bh', 'close']]).\
            reset_index().drop('index', axis=1)

                                        y_pred = model.predict(tf_idf.transform(df_temp[field]))

                                        df_temp['sentiment'] = y_pred

                                        df_temp['neg'] = 0
                                        df_temp['neg'][df_temp.sentiment==-1] = 1

                                        df_temp['pos'] = 0
                                        df_temp['pos'][df_temp.sentiment==1] = 1

                                        df_temp['neut'] = 0
                                        df_temp['neut'][df_temp.sentiment==0] = 1

                                        df_temp = df_temp.drop_duplicates().\
                                        groupby(['date', f'change_close_{day}', 'bh', 'close']).sum().reset_index()

        # Делаем сум роллинг окно для суммирования сентиментов для периодов больше 1 дня 
        # Нужно удостовериться в сортировке должно быть по возрастанию даты

                                        df_temp[f'close_lag_1'] = df_temp.close.shift(periods=1)
                                        #df_temp[f'close_change_lag_1'] = df_temp.close.shift(periods=1)
                                        df_temp = df_temp.sort_values('date', ascending=True)

                                        df_temp['neg'] = df_temp['neg'].rolling(day).sum()
                                        df_temp['pos'] = df_temp['pos'].rolling(day).sum()
                                        df_temp['neut'] = df_temp['neut'].rolling(day).sum()

                                        df_temp.dropna(subset=['neg', 'pos', 'neut', 'close_lag_1'], inplace=True)
                        # просто сумма
                                        df_temp['sum_sentiment'] = 0
                                        df_temp['sum_sentiment'][(df_temp.pos>df_temp.neg)&(df_temp.pos>=df_temp.neut)] = 1
                                        df_temp['sum_sentiment'][(df_temp.pos<df_temp.neg)&(df_temp.neg>=df_temp.neut)] = -1

        # дивергенция c добавлением 
                                        df_temp['diver'] = ((df_temp.pos - df_temp.neg)**2 \
                                                                  / (df_temp.pos + df_temp.neg)**2)**(1/2)

                                        #df_temp.reset_index(level=[f'change_close_{day}', 'date', 'bh'], inplace=True)
                                        df_temp['clear_return'] = df_temp.sum_sentiment * df_temp[f'change_close_{day}']
                                        df_temp['diver_return'] = df_temp.clear_return * df_temp.diver                

                                        res = dict()
                                        res['counter'] = [counter]
                                        res['ticker'] = [ticker]
                                        res['date'] = [day]

                                        res['min_df'] = [min_df]

                                        res['j_neut'] = [j_neut]
                                        res['j_pos'] = [j_pos] 
                                        res['j_neg'] = [j_neg] 

                                        res['train_size'] = [y_train.shape[0]]
                                        res['test_size'] = [y_pred.shape[0]]

                                        res['obs'] = \
                                        [(df_temp.obs[df_temp.clear_return.isna()==False]).sum()]
                                        res['bh'] = df_temp.bh.max()
                                        res['original_return_isna_all'] = \
                                        [(df_temp[f'change_close_{day}']\
                                          [df_temp[f'change_close_{day}'].isna()==False]).sum()]


                                        res['model_return_isna_all'] = \
                                        [(df_temp.clear_return[df_temp.clear_return.isna()==False]).sum()]

                                        res['model_return_isna_short'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_return_isna_long'] = \
                                        [(df_temp.clear_return[(df_temp.clear_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]


                                        res['model_diver_return_isna_all'] = \
                                        [(df_temp.diver_return[df_temp.diver_return.isna()==False]).sum()]

                                        res['model_diver_return_isna_short'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=1)]).sum()]

                                        res['model_diver_return_isna_long'] = \
                                        [(df_temp.diver_return[(df_temp.diver_return.isna()==False) & \
                                                                     (df_temp.sum_sentiment!=-1)]).sum()]






                                        res['train_accuracy'] = [metrics.accuracy_score(y_train_pred, y_train)]


                                        res['train_precision_macro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='macro')]


                                        res['train_precision_micro'] = [metrics.precision_score(y_train_pred, y_train\
                                                                                                , average='micro')]


                                        res['train_recall_micro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='micro')]


                                        res['train_recall_macro'] = [metrics.recall_score(y_train_pred, y_train\
                                                                                          , average='macro')]

                                        res['train_f1_micro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='micro')]


                                        res['train_f1_macro'] = [metrics.f1_score(y_train_pred, y_train\
                                                                                  , average='macro')]



                                        x = df_temp[['sum_sentiment', 'close_lag_1']]
                                        x.sum_sentiment = x.sum_sentiment.astype(str)

                                        x = pd.get_dummies(x, drop_first=True)
                                        y = df_temp.close
                                        x2 = sm.add_constant(x)
                                        lm = sm.OLS(y, x2)
                                        res_lm = lm.fit()

                                        res['const'] = res_lm.params[0]
                                        res[f'{x.columns[0]}_coef'] = res_lm.params[1]
                                        res[f'{x.columns[1]}_coef'] = res_lm.params[2]
                                        res[f'{x.columns[2]}_coef'] = res_lm.params[3]

                                        res['const_p'] = res_lm.pvalues[0]
                                        res[f'{x.columns[0]}_p'] = res_lm.pvalues[1]
                                        res[f'{x.columns[1]}_p'] = res_lm.pvalues[2]
                                        res[f'{x.columns[2]}_p'] = res_lm.pvalues[3]
                                        res['r2'] = res_lm.rsquared
                                        res['nobs'] = res_lm.nobs
                                        counter += 1
                                        print(counter, end='\r')


                                        res = pd.DataFrame.from_dict(res)

                                        out = out.append(pd.DataFrame(res))
                                        if counter % 100 == 0: 
                                            out.to_excel(f'/is/res/nb/mag_nb_model_ind_final1.xlsx', index=False)

                        except:
                            continue
    

                    


out.to_excel(f'/is/res/nb/mag_nb_model_ind_final1.xlsx', index=False)
out
